+++ RELEASENOTES.md
@@ -50,3 +50,4 @@
   - Fix boom generic specials to allow monsters to trigger of off model if change is 0
   - Fix various UMAPINFO issues dealing with nointermission intertext, intertextsecret, endgame, endpic, and intermusic
   - Fix berserk only restoring to initial health when modified through dehacked
+  - Fix OpenGL context using 3.3 with Intel iGPUs when a higher version is supported